renamed button and inputField components and files according to the rules / fixed basic section component/ added example of customizing basic section styles in homePage component
