+++ src/components/InputField/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import * as styles from 'src/components/inputField/inputField.styles';
+import * as styles from 'src/components/inputField/styles';
 import { InputFieldInt } from 'src/components/inputField/types';
 
 const InputField = ({ label, placeholder, validate }: InputFieldInt) => {
