--- src/components/button/buttonEnum.ts
@@ -1,4 +1,0 @@
-export enum ButtonVariant {
-  Black = 'black',
-  White = 'white',
-}
