--- src/components/button/buttonProps.ts
@@ -1,4 +1,0 @@
-export interface buttonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant: string;
-}
